Changed the "run counter" to count all attempted points, not just successful points
This simplified immensely the logic required to avoid race conditions during the last run. It breaks previous scripts (they will not have a guarantee of hitting any points), but it's a good thing for determinism.
I also added a counter of hits per file, which lets you verify at the end of the batch how many hits were actually recorded.